Port client_app to new code and then removed
