+++ appshell/cefclient.h
@@ -8,7 +8,7 @@
 
 #include <string>
 #include "include/cef_base.h"
-#include "client_app.h"
+#include "appshell/common/client_app.h"
 
 class CefApp;
 class CefBrowser;
@@ -43,7 +43,7 @@ void AppInitCommandLine(int argc, const char* const* argv);
 CefRefPtr<CefCommandLine> AppGetCommandLine();
 
 // Returns the application settings based on command line arguments.
-void AppGetSettings(CefSettings& settings, CefRefPtr<ClientApp> app);
+void AppGetSettings(CefSettings& settings, CefRefPtr<client::ClientApp> app);
 
 // Returns the application browser settings based on command line arguments.
 void AppGetBrowserSettings(CefBrowserSettings& settings);
