--- appshell/client_app_delegates.cpp
@@ -1,9 +1,0 @@
-// Copyright (c) 2012 The Chromium Embedded Framework Authors. All rights
-// reserved. Use of this source code is governed by a BSD-style license that
-// can be found in the LICENSE file.
-
-#include "client_app.h"
-
-// static
-void ClientApp::CreateRenderDelegates(RenderDelegateSet& delegates) {
-}
