+++ appshell_paths.gypi
@@ -191,9 +191,6 @@
       'appshell/appshell_node_process.cpp',
       'appshell/command_callbacks.h',
       'appshell/config.h',
-      'appshell/client_app.cpp',
-      'appshell/client_app.h',
-      'appshell/client_app_delegates.cpp',
       'appshell/client_handler.cpp',
       'appshell/client_handler.h',
       'appshell/native_menu_model.cpp',
